Abstract caching from scanManager
